refactor: move run_agent_loop into SubagentPlugin (closes #21)
run_agent_loop only had one caller (SubagentPlugin), but lived in
agent_loop.py alongside run_agent_turn (the shared primitive used by
Agent). The dependency direction was backwards. Reorganize into:

- corvidae/tools/subagent.py — gains run_agent_loop and
  MAX_ROUNDS_REACHED_MESSAGE; now owns the LLM-call-plus-tool-dispatch
  loop next to its only caller.
- corvidae/thinking.py — gains strip_thinking and strip_reasoning_content
  (text utilities operating on LLM output, used by ThinkingPlugin and
  by subagent.py's work() closure).
- corvidae/agent_turn.py — renamed from agent_loop.py; retains only the
  single-LLM-turn primitive (run_agent_turn, AgentTurnResult, _truncate,
  LOG_TRUNCATION_LENGTH).

Tests reorganized to match: run_agent_loop tests moved into
tests/test_subagent.py, strip_* tests moved into tests/test_thinking.py,
and tests/test_agent_loop.py renamed to tests/test_agent_turn.py for
the run_agent_turn tests. Import paths and logger-name caplog filters
updated across all affected test files.

Pure refactor — no behavior changes. All 1046 tests pass; the 9 failures
in test_compaction_config_validation.py and test_idle_monitor.py are
pre-existing on main and unrelated.
